requirements.txt

fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0


langchain==0.0.292
sentence-transformers==2.2.2
faiss-cpu>=1.7.0
numpy==1.25.2
pandas==2.1.0


transformers==4.36.0
accelerate==0.25.0
bitsandbytes==0.41.0
torch==2.0.1
sentencepiece==0.1.99

